Start acting on some of the things I have reported bugs on (#8)
* Format with toml-formatter

* Update required python to be the same as llm

Doesn't make sense to support older python versions than the plugin
host. Also everything before 3.9 is already end of lifed.

* Add watching-testrunner to make running tests simpler

This allows 'watching_testrunner pytest' which runs the testsuite on
save in the editor making the test -> feedback loop that much shorter.

* Ensure uv installs the package when syncing

Because the pyproject.toml does not specify a build system, uv assumes
that it should not be installed. Not helpfull.

* Add a lockfile, to ensure reproducible failures

* Switch to modern type annotations

* Update type annotations

And remove obsoleted less precise code comments

* Simplify mocking

* Switch to pathlib

* Ignore pyright config